requirements.txt
sqlalchemy==1.3.9
psycopg2==2.9.3
python-dotenv==0.20.0
